backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
passlib[bcrypt]
jwt
torch
torchvision
numpy
pillow
